add db

+++ db/migrate/20160528132228_create_user_items.rb
@@ -2,15 +2,15 @@ class CreateUserItems < ActiveRecord::Migration
   def change
     create_table :user_items do |t|
       t.string :user_id
-      t.integer :first_1_sum
-      t.integer :first_2_sum
-      t.integer :second_1_sum
-      t.integer :second_2_sum
-      t.integer :third_1_sum
-      t.integer :third_2_sum
-      t.integer :fourth_1_sum
-      t.integer :fourth_2_sum
-      t.integer :sum_sum
+      t.float :first_1_sum
+      t.float :first_2_sum
+      t.float :second_1_sum
+      t.float :second_2_sum
+      t.float :third_1_sum
+      t.float :third_2_sum
+      t.float :fourth_1_sum
+      t.float :fourth_2_sum
+      t.float :sum_sum
       t.string :first_1_flg
       t.string :first_2_flg
       t.string :second_1_flg
